Divide codes into modules and used ES6 syntax
