+++ .gitignore
@@ -1,0 +1,3 @@
+.vscode/
+
+node_modules/
