+++ index.html
@@ -12,40 +12,40 @@
         <h1>Awesome Books</h1>
         <nav>
           <ul class="links">
-            <li><a href="#" class="navlinks first">List</a></li>
-            <li><a href="#" class="navlinks">Add New</a></li>
-            <li><a href="#" class="navlinks">Contact</a></li>
+            <li><a id="headlist" href="#" class="active">List</a></li>
+            <li><a id="headadd" href="#">Add New</a></li>
+            <li><a id="headcontact" href="#">Contact</a></li>
           </ul>
         </nav>
     </div>
 
-    <section>
-      <h2>All Awesome Books</h2>
-      <ul class="bookstack"></ul>
-    </section>
-
-    <section>
-      <h2>Add a new book</h2>
+    <section id="liststack">
+      <div class="book-container">
+      <h2 class="stacks">All Awesome Books</h2>
       <div class="booklist">
       </div>
-  
-      <form id="book-form">
-        <div>
-        <input type="text" id="booktitle" name="title" placeholder="Title">
-        </div>
-       <div>
-        <input type="text" id="bookauthor" name="author" placeholder="Author">
-        </div>
+      </div>
+      <!-- <ul class="bookstack"></ul> -->
+    </section>
 
+    <section id="addbook-section" class="hide-content">
+      <div class="form-container">
+      <h2>Add a new book</h2>
+      
+      <form id="book-form">
+        <input type="text" id="booktitle" name="title" placeholder="Title">
+        <input type="text" id="bookauthor" name="author" placeholder="Author">
         <div class="btn">
         <button id="addbtn" type="submit">Add</button>
         </div>
     </form>
+    </div>
     
   </section>
 
-  <section class="contact">
-    <h2>Contact Information</h2>
+  <section id="contact" class="hide-content">
+    <div class="contact-info">
+    <h2 class="reach">Contact Information</h2>
     <p>
       Do you have any question or you just want to say "Hello"? 
       You can reach out to us!
@@ -55,13 +55,13 @@
       <li class="info">Our phone number: 0043586534422</li>
       <li class="info">Our address: Streetname 22, 84503 city, country</li>
     </ul>
-
+    </div>
   </section>
 
   <footer>
-    Copyright 2021
+    <p>Copyright 2021</p>
   </footer>
 
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
